fix: resolve package map PR checks
