+++ .changeset/fix-yarn-import-js-yaml.md
@@ -1,0 +1,6 @@
+---
+"@pnpm/installing.commands": patch
+"pnpm": patch
+---
+
+Fixed `pnpm import` for Yarn v2 lockfiles when `js-yaml` v4 is installed.
