+++ installing/commands/package.json
@@ -85,7 +85,6 @@
     "@pnpm/workspace.workspace-manifest-writer": "workspace:*",
     "@yarnpkg/core": "catalog:",
     "@yarnpkg/lockfile": "catalog:",
-    "@yarnpkg/parsers": "catalog:",
     "@zkochan/rimraf": "catalog:",
     "@zkochan/table": "catalog:",
     "chalk": "catalog:",
@@ -93,6 +92,7 @@
     "detect-libc": "catalog:",
     "get-npm-tarball-url": "catalog:",
     "is-subdir": "catalog:",
+    "js-yaml": "catalog:",
     "load-json-file": "catalog:",
     "normalize-path": "catalog:",
     "p-filter": "catalog:",
@@ -119,6 +119,7 @@
     "@pnpm/testing.registry-mock": "workspace:*",
     "@pnpm/worker": "workspace:*",
     "@pnpm/workspace.projects-filter": "workspace:*",
+    "@types/js-yaml": "catalog:",
     "@types/normalize-path": "catalog:",
     "@types/proxyquire": "catalog:",
     "@types/ramda": "catalog:",
